+++ src/rti1516e/RTI1516ELogicalTimeFactory.h
@@ -1,4 +1,4 @@
-/* -*-c++-*- OpenRTI - Copyright (C) 2009-2015 Mathias Froehlich
+/* -*-c++-*- OpenRTI - Copyright (C) 2009-2016 Mathias Froehlich
  *
  * This file is part of OpenRTI.
  *
@@ -117,7 +117,7 @@ public:
   LogicalTime decodeLogicalTime(const VariableLengthData& variableLengthData);
   VariableLengthData encodeLogicalTime(const rti1516e::LogicalTime& logicalTime)
   {
-    unsigned long encodedLength = logicalTime.encodedLength();
+    size_t encodedLength = logicalTime.encodedLength();
     VariableLengthData variableLengthData;
     variableLengthData.resize(encodedLength);
     logicalTime.encode(variableLengthData.data(), encodedLength);
